requirements.txt
fastapi
pydantic
langchain
langchain-openai
langchain-google-genai
qdrant-client
python-dotenv
numpy
streamlit
